-water progress indicator fixed

+++ app/src/main/java/com/johnzieman/ziemapp/drinkwater/models/WaterDaily.kt
@@ -8,8 +8,8 @@ import java.util.*
 data class WaterDaily(
     @PrimaryKey
     val id: UUID = UUID.randomUUID(),
-    val dailyRate: Int,
-    var drunk: Int,
+    val dailyRate: Float,
+    var drunk: Float,
     var cupsRate: Int,
     var cupDrunk: Int
 )
